server/requirements.txt
blinker==1.9.0
click==8.1.7
Flask==3.0.3
Flask-Cors==5.0.0
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
PyMySQL==1.1.1
python-dotenv==1.0.1
pytz==2024.2
typing_extensions==4.12.2
Werkzeug==3.1.3
zipp==3.21.0
cryptography==43.0.3

